# src/routes.rs
use std::sync::Arc;

use axum::{
    Router, async_trait,
    extract::{FromRequest, Request, rejection::JsonRejection},
    http::StatusCode,
    middleware,
    response::{IntoResponse, Response},
    routing::{get, post},
};
use serde::Serialize;
use serde_json::{Value, json};
use tower_http::cors::CorsLayer;

use crate::{
    routes::{
        control_request::{handle_reset, handle_start},
        status::{stats_handler, status_handler},
        ws::ws_handler,
    },
    state::ApiState,
};

use self::{update_output::handle_keyframe_request, update_output::handle_output_update};
use crate::middleware::body_logger_middleware;

pub mod control_request;
pub mod register_request;
pub mod status;
pub mod unregister_request;
pub mod update_input;
pub mod update_output;
pub mod ws;

pub fn routes(state: Arc<ApiState>) -> Router {
    let inputs = Router::new()
        .route("/:id/register", post(register_request::handle_input))
        .route("/:id/unregister", post(unregister_request::handle_input))
        .route("/:id/update", post(update_input::handle_input_update));

    let outputs = Router::new()
        .route("/:id/register", post(register_request::handle_output))
        .route("/:id/unregister", post(unregister_request::handle_output))
        .route("/:id/update", post(handle_output_update))
        .route("/:id/request_keyframe", post(handle_keyframe_request));

    let image = Router::new()
        .route("/:id/register", post(register_request::handle_image))
        .route("/:id/unregister", post(unregister_request::handle_image));

    let web = Router::new()
        .route("/:id/register", post(register_request::handle_web_renderer))
        .route(
            "/:id/unregister",
            post(unregister_request::handle_web_renderer),
        );

    let shader = Router::new()
        .route("/:id/register", post(register_request::handle_shader))
        .route("/:id/unregister", post(unregister_request::handle_shader));

    let font = Router::new().route("/register", post(register_request::handle_font));

    Router::new()
        .nest("/api/input", inputs)
        .nest("/api/output", outputs)
        .nest("/api/image", image)
        .nest("/api/web-renderer", web)
        .nest("/api/shader", shader)
        .nest("/api/font", font)
        // Start request
        .route("/api/start", post(handle_start))
        .route("/api/reset", post(handle_reset))
        // WebSocket - events
        .route("/ws", get(ws_handler))
        .route("/status", get(status_handler))
        .route("/stats", get(stats_handler))
        .layer(CorsLayer::permissive())
        .layer(middleware::from_fn(body_logger_middleware))
        .with_state(state)
}

/// Wrap axum::Json to return serialization errors as json
pub struct Json<T>(pub T);

impl<T: Serialize> IntoResponse for Json<T> {
    fn into_response(self) -> Response {
        axum::Json(self.0).into_response()
    }
}

#[async_trait]
impl<S, T> FromRequest<S> for Json<T>
where
    axum::Json<T>: FromRequest<S, Rejection = JsonRejection>,
    S: Send + Sync,
{
    type Rejection = (StatusCode, axum::Json<Value>);

    async fn from_request(req: Request, state: &S) -> Result<Self, Self::Rejection> {
        match axum::Json::<T>::from_request(req, state).await {
            Ok(value) => Ok(Self(value.0)),
            Err(rejection) => {
                let payload = json!({
                    "error_code": "MALFORMED_REQUEST",
                    "message": rejection.body_text(),
                });

                Err((rejection.status(), axum::Json(payload)))
            }
        }
    }
}

pub struct Multipart(pub axum::extract::Multipart);

#[async_trait]
impl<S> FromRequest<S> for Multipart
where
    S: Send + Sync,
{
    type Rejection = (StatusCode, axum::Json<Value>);

    async fn from_request(req: Request, state: &S) -> Result<Self, Self::Rejection> {
        match axum::extract::Multipart::from_request(req, state).await {
            Ok(multipart) => Ok(Multipart(multipart)),
            Err(rejection) => {
                let payload = json!({
                    "error_code": "MALFORMED_MULTIPART",
                    "message": rejection.body_text(),
                });

                Err((StatusCode::BAD_REQUEST, axum::Json(payload)))
            }
        }
    }
}