# tools/schemas/openapi_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smelter",
    "description": "Real-time video compositing software",
    "license": {
      "name": "MIT"
    },
    "version": "0.6.0"
  },
  "paths": {
    "/api/start": {
      "post": {
        "tags": [
          "control_request"
        ],
        "operationId": "start",
        "responses": {
          "200": {
            "description": "Smelter instance started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reset": {
      "post": {
        "tags": [
          "control_request"
        ],
        "operationId": "reset",
        "responses": {
          "200": {
            "description": "Smelter instance reset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/input/{input_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_input",
        "parameters": [
          {
            "name": "input_id",
            "in": "path",
            "description": "Input ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Input registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterInputResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterOutput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOutputResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shader/{shader_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_shader",
        "parameters": [
          {
            "name": "shader_id",
            "in": "path",
            "description": "Shader ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShaderSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shader registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/web-renderer/{instance_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_web_renderer",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Web renderer instance ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebRendererSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Web renderer registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image/{image_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/font/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_font",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFontRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Font registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/input/{input_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_input",
        "parameters": [
          {
            "name": "input_id",
            "in": "path",
            "description": "Input ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Input unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Input not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Output not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shader/{shader_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_shader",
        "parameters": [
          {
            "name": "shader_id",
            "in": "path",
            "description": "Shader ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shader unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Shader not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/web-renderer/{instance_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_web_renderer",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Web renderer ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Web renderer unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Web renderer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image/{image_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Image not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/input/{input_id}/update": {
      "post": {
        "tags": [
          "update_request"
        ],
        "operationId": "update_input",
        "parameters": [
          {
            "name": "input_id",
            "in": "path",
            "description": "Input ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Input updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Input not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/update": {
      "post": {
        "tags": [
          "update_request"
        ],
        "operationId": "update_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/request_keyframe": {
      "post": {
        "tags": [
          "update_request"
        ],
        "operationId": "request_keyframe",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keyframe request successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "tags": [
          "metadata_request"
        ],
        "operationId": "get_status",
        "responses": {
          "200": {
            "description": "Instance status fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceStatus"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "metadata_request"
        ],
        "operationId": "get_stats",
        "responses": {
          "200": {
            "description": "Statistics for inputs and outputs fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsReport"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/ws": {
      "get": {
        "tags": [
          "ws_request"
        ],
        "operationId": "ws",
        "responses": {
          "200": {
            "description": "WebSocket connection started succesfully."
          },
          "400": {
            "description": "Bad request."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "stack",
          "http_status_code"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "http_status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Response for requests that do not return any data."
      },
...
      "RegisterFontRequest": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "contentMediaType": "application/octet-stream"
          }
        }
      },
...
      "RegisterInputResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "port": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "properties": {
              "video_duration_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              },
              "audio_duration_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "required": [
              "bearer_token",
              "endpoint_route"
            ],
            "properties": {
              "bearer_token": {
                "type": "string"
              },
              "endpoint_route": {
                "type": "string"
              }
            }
          },
          {
            "type": "object"
          }
        ]
      },
...
      "RegisterOutputResponse": {
        "type": "object",
        "properties": {
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Port allocated for the output. Only returned for RTP outputs.",
            "minimum": 0
          }
        }
      },
...
      "ShaderSpec": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "Shader source code."
          }
        },
        "additionalProperties": false
      },
...
      "UnregisterRequest": {
        "type": "object",
        "description": "Request body shared by all unregister routes.",
        "properties": {
          "schedule_time_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Time in milliseconds when this request should be applied. Value `0` represents\ntime of the start request."
          }
        }
      },
      "UpdateInputRequest": {
        "type": "object",
        "properties": {
          "pause": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "seek_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Seek to a specific position in milliseconds. Only supported for MP4 inputs."
          }
        },
        "additionalProperties": false
      },
...
      "WebEmbeddingMethod": {
        "type": "string",
        "enum": [
          "chromium_embedding",
          "native_embedding_over_content",
          "native_embedding_under_content"
        ]
      },
      "WebRendererSpec": {
        "type": "object",
        "required": [
          "url",
          "resolution"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "Url of a website that you want to render."
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution",
            "description": "Resolution."
          },
          "embedding_method": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WebEmbeddingMethod",
                "description": "Mechanism used to render input frames on the website."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}